fix(partial): resolving clippy lints
Removing the `#[warn(dead_code)]` flag revealed several unused funtions
that probably _should_ be used in some way.

1. The current state is a one-leg-in, one-leg-out use of the
`ContainerState` struct.

There was a shift during development to use docker as the source of
truth for state rather than rely on a stateful map tracking which
containers belong to the launchpad and their state.

However, this has not been fully completed. There's still a reliance on
a `CONTAINERS` hashmap. To fully migrate to using the docker deamon as
the source of truth, this needs to be deprcated so that container
management is stateless. We need

- a convention for identifying containers that "haz belong to us". This
  is most likely a pure naming convention.
- adding, listing, filtering, and removing containers must all be
  stateless, query the docker API directly, and not rely (or as is currently, half-rely) on a
CONTAINERS hashmap.

2. There are functions for hitting docker.com and quay.io to check if
   new images are present, but the latter don't appear to be wired up to
the main app??!!

3. I've removed all tests that actually use docker since they were
   hyper-flaky and fragile. But clearly tests are still needed.