Comment out image update status
Fetching image statuses works fine but fetching image information and
working out if an image needs updating is broken at every step. The two
apis we need to communicate with will return data in different formats.
Docker Hub also doesn't return image SHA's the will match the local
sha's of the exact same image. Multi platofrm images provide an extra
layer of complexity that will need consideration when comparing the
images.

All of this will require a well thought through process and isn't as
easy as a bug fix. In the meantime turn we've surrended and decided to
simply turn off update notifications for now. It's a band-aid, but a
better choice over the rabbit hole.