feat: randomise tor control password
The tor control port password can be made ephemeral and random when the
tor container is started. This password can be supplied to other
containers so that they can access Tor.

This prevents actors that have access to the docker daemon from using
the Tor control port -- assuming they're unable to `exec` or `attach` to
any running container in order to access the environment.