fix: startup issues
When starting from a completely clean build, the node doesn't create the
id.

This requires running in non-interactive mode for now.

There's also a bug in the comms layer where the wallet's peer db must be deleted
in order to get the wallet to connect internally to get the wallet to
reconnect via TCP on next launch.