feat: fix startup bugs (#265)
1. Wallet

There are 2 extra settings in the wallet config. clear_peer_db, clears
the peer database when configuring the wallet.
This should be true for at least the first run or until
https://github.com/tari-project/tari/issues/5998 is fixed.

`interacive` must be false for the first run, and can be set to true
after that.
Adds a quick fix for the wallet for the non-interactive and peer DB
We only display the instructions for attaching to the docker wallet
instance if interactive is true, and the wallet is active.

bugs.

You can control whether the wallet is non-interactive (default. will run
on first startup), and whether the peer DB is deleted (true. shold let
2. Base node

Similarly to the wallet, there's an `interactive` setting for the base
node now that must be false on the first run.

A future PR should make these editable in the UI.